code review v3 - POO

+++ src/Controllers/PostsController.php
@@ -3,6 +3,7 @@
 namespace App\controllers;
 
 use App\Model\Repository\PostRepository;
+use App\Model\Repository\CommentRepository;
 use App\Model\Repository\UserRepository;
 use App\lib\DatabaseConnection;
 use App\Model\Entity\Post;
@@ -10,12 +11,14 @@ use App\Model\Entity\Post;
 class PostsController
 {
     private $postRepository;
+    private $commentRepository;
     private $userRepository;
     private $currentTime;
 
     public function __construct(DatabaseConnection $connection)
     {
         $this->postRepository = new PostRepository($connection);
+        $this->commentRepository = new CommentRepository($connection);
         $this->userRepository = new UserRepository($connection);
         $this->currentTime = date('Y-m-d H:i:s');
     }
@@ -23,12 +26,13 @@ class PostsController
 
     public function getPosts()
     {
-        $allPosts = $this->postRepository->getAllPosts();
+        $posts = $this->postRepository->getAllPosts();
         require_once(__DIR__.'/../../templates/admin/admin_posts_list_page.php');
     }
 
     public function getPublishedPosts()
     {
+
         $posts = $this->postRepository->getPublishedPosts();
         require(__DIR__.'/../../templates/posts_list_page.php');
     }
@@ -36,6 +40,7 @@ class PostsController
     public function getPostById($postId)
     {
         $post = $this->postRepository->getPostById($postId);
+        $comments = $this->commentRepository->getValidComments($post->id);
         require(__DIR__.'/../../templates/show_post_page.php');
     }
 
@@ -45,25 +50,24 @@ class PostsController
         require(__DIR__.'/../../templates/admin/posts/add_post_page.php');
     }
 
-    public function postAddPost($user_id, $title, $chapo, $content, $is_published)
+    public function postAddPost($data)
     {
-        var_dump($title, $content);
-        if (!isset($title) || !isset($content)) {
+        if (!isset($data["title"]) || !isset($data["content"])) {
             echo "Veuillez remplir les champs Titre et Contenu.";
             return;
         }
 
-        $user = $this->userRepository->getUserById($user_id);
-        $user_surname = $user['surname'];
+
+        $user = $this->userRepository->getUserById($data["user_id"]);
 
         $post = new Post();
-        $post->setTitle($title);
-        $post->setChapo($chapo);
-        $post->setContent($content);
-        $post->setAuthor($user_id);
-        $post->setIsPublished($is_published);
-        $post->setCreatedAt($this->currentTime);
-        $post->setUpdatedAt($this->currentTime);
+        $post->init(
+            $data["title"],
+            $data["chapo"],
+            $data["content"],
+            intval($data["user_id"]),
+            isset($data["is_published"]) ? 1 : 0
+        );
 
         $this->postRepository->addPost($post);
 
@@ -74,20 +78,28 @@ class PostsController
     public function getEditPost($postId)
     {
         $post = $this->postRepository->getPostById($postId);
+        $admins = $this->userRepository->getAdmins();
         require(__DIR__.'/../../templates/admin/posts/edit_post_page.php');
     }
 
-    public function postEditPost($postId, $user_id, $title, $chapo, $content, $is_published)
+    public function postEditPost($data)
     {
-        $post = $this->postRepository->getPostById($postId);
 
-        if($post) {
-            $post->setTitle($title);
-            $post->setChapo($chapo);
-            $post->setContent($content);
-            $post->setAuthor($user_id);
-            $post->setIsPublished($is_published);
-            $post->setUpdatedAt($this->currentTime);
+        if (!isset($data["post_id"]) || !isset($data["user_id"]) || !isset($data["title"]) || !isset($data["content"])) {
+            echo "Veuillez fournir toutes les informations nécessaires.";
+            return;
+        }
+
+        $post = $this->postRepository->getPostById($data["post_id"]);
+
+        if ($post) {
+            $post->update(
+                $data["title"],
+                $data["chapo"],
+                $data["content"],
+                $data["user_id"],
+                isset($data["is_published"]) ? 1 : 0
+            );
 
             $this->postRepository->editPost($post);
 
@@ -103,14 +115,17 @@ class PostsController
         require(__DIR__.'/../../templates/admin/posts/delete_post_page.php');
     }
 
-    public function postDeletePost($postId)
+    public function postDeletePost($data)
     {
-        $post = $this->postRepository->getPostById($postId);
+        if (!isset($data["post_id"])) {
+            echo "Veuillez fournir l'identifiant du post.";
+            return;
+        }
+
+        $post = $this->postRepository->getPostById($data["post_id"]);
 
         if($post) {
-            $post->setIsPublished(0);
-            $post->setUpdatedAt($this->currentTime);
-            $post->setDeletedAt($this->currentTime);
+            $post->delete($data["post_id"], 0);
 
             $this->postRepository->deletePost($post);
 
@@ -127,13 +142,12 @@ class PostsController
         require(__DIR__.'/../../templates/admin/posts/restore_post_page.php');
     }
 
-    public function postRestorePost($postId)
+    public function postRestorePost($data)
     {
-        $post = $this->postRepository->getPostById($postId);
+        $post = $this->postRepository->getPostById($data["post_id"]);
 
         if($post) {
-            $post->setDeletedAt(null);
-            $post->setUpdatedAt($this->currentTime);
+            $post->restore($data["post_id"]);
 
             $this->postRepository->restorePost($post);
